[server] roze/stream rename to flow metrics/flow log
