+++ server/ingester/ingesterctl/const.go
@@ -30,8 +30,8 @@ const (
 	INGESTERCTL_RPC
 	INGESTERCTL_LOGLEVEL
 	INGESTERCTL_CONFIG
-	INGESTERCTL_ROZE_QUEUE
-	INGESTERCTL_STREAM_QUEUE
+	INGESTERCTL_FLOW_METRICS_QUEUE
+	INGESTERCTL_FLOW_LOG_QUEUE
 	INGESTERCTL_EXTMETRICS_QUEUE
 	INGESTERCTL_PCAP_QUEUE
