--- server/ingester/pkg/influxdb-migrate/influxdb_migrate.yaml
@@ -1,80 +1,0 @@
-## 用来校验升级路径是否正确
-migrate-path: 570-571
-
-# droplet-config: /etc/droplet.yaml
-
-# shard-id: 0
-# tsdb-src:
-#   addr: http://127.0.0.1:20044
-#   user: root
-#   password: root
-
-# ck-addr:
-#   addr: tcp://127.0.0.1:9000
-#   user: default
-#   password: root
-
-
-## 并行处理数目
-# parallel-count: 4
-
-## 限制写入速度, 目前同时写主备，如果多个分析器同时迁移,同时还有新数据写入. 故当个分析器默认限制100000条每秒
-limit-write-rate: 100000
-
-migrates-1m:
-- dbs:
-  - vtap_flow 
-  - vtap_flow_edge
-  - vtap_flow_port
-  - vtap_flow_edge_port
-  time-range: ""       #迁移的时间范围条件格式:  "time > now()-1d and time < now()" 默认:""是所有
-- dbs:
-  - vtap_packet
-  - vtap_packet_edge
-  time-range: ""
-- dbs:
-  - vtap_wan
-  - vtap_wan_port
-  time-range: ""
-- dbs:
-  - vtap_acl
-  time-range: ""
-
-# # 自定义数据源的迁移，需要根据局点情况，具体配置
-#migrates-1h1d:
-#- dbs:
-#  - vtap_flow
-#  - vtap_flow_edge
-#  - vtap_flow_port
-#  - vtap_flow_edge_port
-#  rp: 1h      # 名称是  web界面显示的名称, 实际底层存储会使用rp_1h
-#  ck-table: 1h # 一般和rp相同，除非要迁移改名
-#  unsummable: avg  # 非累加聚合的值 avg, max, min
-#  interval: 1h   # 聚合的时间粒度 1h,1d
-#  time-range: "time<now()-7d" # 需要排除分钟数据的时间
-#- dbs:
-#  - vtap_packet
-#  - vtap_packet_edge
-#  rp: 1h
-#  ck-table: 1h
-#  unsummable: avg
-#  interval: 1h
-#  time-range: "time<now()-7d" # 需要排除分钟数据的时间
-#- dbs:
-#  - vtap_wan
-#  - vtap_wan_port
-#  rp: 1h
-#  ck-table: 1h
-#  unsummable: avg
-#  interval: 1h
-#  time-range: "time<now()-7d" # 需要排除分钟数据的时间
-#- dbs:
-#  - vtap_flow
-#  - vtap_flow_edge
-#  - vtap_flow_port
-#  - vtap_flow_edge_port
-#  rp: 1d      # 名称是  web界面显示的名称
-#  ck-table: 1d
-#  unsummable: avg  # 非累加聚合的值 avg, max, min
-#  interval: 1d   # 聚合的时间粒度 1h,1d
-#  time-range: "time<now()-30d" # 需要排除小时或分钟数据的时间
